chore : 코딩 리뷰 1차 수정
